--- L24-webServer/src/main/java/ru/otus/WebServerSimpleDemo.java
@@ -1,41 +1,0 @@
-package ru.otus;
-
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import ru.otus.crm.service.DBServiceClient;
-import ru.otus.crm.service.DbServiceClientImpl;
-import ru.otus.dao.InMemoryUserDao;
-import ru.otus.dao.UserDao;
-import ru.otus.server.UsersWebServer;
-import ru.otus.server.UsersWebServerSimple;
-import ru.otus.services.*;
-
-/*
-    Полезные для демо ссылки
-
-    // Стартовая страница
-    http://localhost:8080
-
-    // Страница пользователей
-    http://localhost:8080/users
-
-    // REST сервис
-    http://localhost:8080/api/user/3
-*/
-public class WebServerSimpleDemo {
-    private static final int WEB_SERVER_PORT = 8080;
-    private static final String TEMPLATES_DIR = "/templates/";
-
-    public static void main(String[] args) throws Exception {
-        UserDao userDao = new InMemoryUserDao();
-        Gson gson = new GsonBuilder().serializeNulls().setPrettyPrinting().create();
-        TemplateProcessor templateProcessor = new TemplateProcessorImpl(TEMPLATES_DIR);
-        DBServiceClient dbServiceClient = DbServiceClientImpl.getDBServiceClient();
-
-        UsersWebServer usersWebServer = new UsersWebServerSimple(WEB_SERVER_PORT, userDao,
-                gson, templateProcessor, dbServiceClient);
-
-        usersWebServer.start();
-        usersWebServer.join();
-    }
-}
